refactor(types): move modifier out from keycode mod
