+++ rmk-types/src/action.rs
@@ -1,4 +1,5 @@
-use crate::keycode::{KeyCode, modifier::ModifierCombination};
+use crate::keycode::KeyCode;
+use crate::modifier::ModifierCombination;
 
 /// EncoderAction is the action at a encoder position, stored in encoder_map.
 #[derive(Clone, Copy, Debug)]
