+++ rmk-types/src/lib.rs
@@ -3,4 +3,5 @@
 pub mod action;
 pub mod hid_state;
 pub mod keycode;
+pub mod modifier;
 pub mod protocol;
